fix build issues on linux/mac

+++ shared/ebm_native/ebm_native_test/EbmNativeTest.cpp
@@ -746,7 +746,7 @@ FloatEbmType TestApi::Boost(
       // if sums are on, then we MUST change the model update
 
       size_t cValues = GetVectorLength(m_learningTypeOrCountTargetClasses);
-      std::vector<size_t> & countBinsByFeatureGroup = m_countBinsByFeatureGroup[indexFeatureGroup];
+      std::vector<size_t> & countBinsByFeatureGroup = m_countBinsByFeatureGroup[static_cast<size_t>(indexFeatureGroup)];
 
       for(size_t iDimension = 0; iDimension < countBinsByFeatureGroup.size(); ++iDimension) {
          size_t cBins = countBinsByFeatureGroup[iDimension];
